refactor(ui): split shared icon modules
