+++ apps/packages/ui/src/icons/types.ts
@@ -1,0 +1,3 @@
+import type { SVGProps } from "react";
+
+export type IconProps = SVGProps<SVGSVGElement>;
