+++ apps/packages/ui/src/icons/workspace.tsx
@@ -1,0 +1,245 @@
+import type { IconProps } from "./types";
+
+export function GitMerge({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <circle cx="18" cy="18" r="3" />
+      <circle cx="6" cy="6" r="3" />
+      <path d="M6 21V9a9 9 0 0 0 9 9" />
+    </svg>
+  );
+}
+
+export function GitBranch({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <line x1="6" x2="6" y1="3" y2="15" />
+      <circle cx="18" cy="6" r="3" />
+      <circle cx="6" cy="18" r="3" />
+      <path d="M18 9a9 9 0 0 1-9 9" />
+    </svg>
+  );
+}
+
+export function GitPullRequest({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <circle cx="18" cy="18" r="3" />
+      <circle cx="6" cy="6" r="3" />
+      <path d="M13 6h3a2 2 0 0 1 2 2v7" />
+      <line x1="6" x2="6" y1="9" y2="21" />
+    </svg>
+  );
+}
+
+export function Terminal({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <polyline points="4 17 10 11 4 5" />
+      <line x1="12" y1="19" x2="20" y2="19" />
+    </svg>
+  );
+}
+
+export function FileText({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" />
+      <path d="M14 2v4a2 2 0 0 0 2 2h4" />
+      <path d="M10 13H8" />
+      <path d="M16 17H8" />
+      <path d="M16 13h-2" />
+    </svg>
+  );
+}
+
+export function FileIcon({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" />
+      <path d="M14 2v4a2 2 0 0 0 2 2h4" />
+    </svg>
+  );
+}
+
+export function InlinePathMentionIcon({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} width="10" height="10" viewBox="0 0 10 10" fill="none" {...props}>
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M5.62988 1.12599C5.9198 1.12599 6.11903 1.12407 6.31006 1.16993L6.42969 1.20362C6.54783 1.24203 6.66141 1.29433 6.76758 1.35939L6.8291 1.39943C6.97079 1.49824 7.09992 1.62972 7.2793 1.80909L7.77442 2.30421L7.91651 2.44728C8.04775 2.58071 8.14716 2.69039 8.22412 2.81593L8.28516 2.92482C8.34146 3.0354 8.38453 3.1525 8.41358 3.27345L8.42871 3.34571C8.459 3.51573 8.45752 3.70002 8.45752 3.95362V6.542C8.45752 6.88641 8.45784 7.16509 8.43945 7.39015C8.42307 7.59041 8.39048 7.77088 8.31836 7.93898L8.28516 8.01027C8.15244 8.27068 7.95039 8.48862 7.70264 8.64064L7.59326 8.70167C7.40494 8.7976 7.20206 8.83726 6.97315 8.85597C6.74803 8.87436 6.46954 8.87452 6.125 8.87452H3.875C3.53046 8.87452 3.25197 8.87436 3.02686 8.85597C2.82659 8.8396 2.64613 8.80745 2.47803 8.73536L2.40674 8.70167C2.14617 8.56891 1.92793 8.36707 1.77588 8.11915L1.71484 8.01027C1.61894 7.822 1.57927 7.61897 1.56055 7.39015C1.54216 7.16509 1.54248 6.88641 1.54248 6.542V3.45851C1.54248 3.11403 1.54217 2.83546 1.56055 2.61036C1.57925 2.38151 1.61898 2.17852 1.71484 1.99025C1.86655 1.6925 2.109 1.45007 2.40674 1.29835C2.59504 1.20245 2.79796 1.16276 3.02686 1.14405C3.25198 1.12566 3.53045 1.12599 3.875 1.12599H5.62988ZM3.875 1.79103C3.51948 1.79103 3.27281 1.79147 3.08106 1.80714C2.89321 1.82249 2.7875 1.85087 2.7085 1.89112C2.5359 1.97909 2.39557 2.1194 2.30762 2.292C2.26739 2.37099 2.23898 2.4768 2.22363 2.66456C2.20798 2.8563 2.20752 3.10308 2.20752 3.45851V6.542C2.20752 6.89736 2.20797 7.14424 2.22363 7.33595C2.23899 7.52361 2.26738 7.62955 2.30762 7.70851L2.34277 7.7715C2.43093 7.91522 2.55744 8.03242 2.7085 8.10939L2.77344 8.13722C2.84532 8.16295 2.94008 8.18185 3.08106 8.19337C3.27281 8.20904 3.51949 8.20948 3.875 8.20948H6.125C6.48051 8.20948 6.72719 8.20904 6.91895 8.19337C7.10673 8.17803 7.21251 8.14961 7.29151 8.10939L7.35449 8.07374C7.49817 7.98564 7.6154 7.85948 7.69238 7.70851L7.7207 7.64308C7.74635 7.57129 7.76487 7.47652 7.77637 7.33595C7.79203 7.14424 7.79248 6.89736 7.79248 6.542V4.27882L6.67529 4.1548C6.12859 4.09405 5.69878 3.65917 5.64404 3.11183L5.51172 1.79103H3.875ZM6.30567 3.04591C6.32918 3.281 6.51374 3.46752 6.74854 3.49366L7.78809 3.60939C7.78635 3.56879 7.7843 3.53557 7.78076 3.50636L7.76709 3.42872C7.75025 3.35858 7.72504 3.29071 7.69238 3.22657L7.65723 3.16359C7.63125 3.1212 7.59976 3.0807 7.54639 3.02247L7.3042 2.77443L6.80908 2.27931C6.63847 2.1087 6.55244 2.02455 6.48438 1.9712L6.41992 1.92628C6.35836 1.88856 6.29263 1.85821 6.22412 1.83595L6.18359 1.82423L6.30567 3.04591Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
+
+export function FileCode({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" />
+      <path d="M14 2v4a2 2 0 0 0 2 2h4" />
+      <path d="m10 13-2 2 2 2" />
+      <path d="m14 17 2-2-2-2" />
+    </svg>
+  );
+}
+
+export function StackedFiles({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth={1.5} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M4.5 2.25h4.25l3.75 3.75v7a.75.75 0 0 1-.75.75H4.5a1 1 0 0 1-1-1v-9.5a1 1 0 0 1 1-1Z" />
+      <path d="M8.75 2.25v3.5a.25.25 0 0 0 .25.25h3.5" />
+    </svg>
+  );
+}
+
+export function FilePen({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M12 22h6a2 2 0 0 0 2-2V7l-5-5H6a2 2 0 0 0-2 2v10" />
+      <path d="M14 2v4a2 2 0 0 0 2 2h4" />
+      <path d="M10.4 12.6a2 2 0 1 1 3 3L8 21l-4 1 1-4Z" />
+    </svg>
+  );
+}
+
+export function FilePlus({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" />
+      <path d="M14 2v4a2 2 0 0 0 2 2h4" />
+      <path d="M9 15h6" />
+      <path d="M12 18v-6" />
+    </svg>
+  );
+}
+
+export function FolderPlus({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M20 20a2 2 0 0 0 2-2V8a2 2 0 0 0-2-2h-7.9a2 2 0 0 1-1.69-.9L9.6 3.9A2 2 0 0 0 7.93 3H4a2 2 0 0 0-2 2v13a2 2 0 0 0 2 2Z" />
+      <line x1="12" x2="12" y1="10" y2="16" />
+      <line x1="9" x2="15" y1="13" y2="13" />
+    </svg>
+  );
+}
+
+export function FolderList({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.75} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M3 7.5A2.5 2.5 0 0 1 5.5 5H9l1.7 2H18.5A2.5 2.5 0 0 1 21 9.5v8A2.5 2.5 0 0 1 18.5 20h-13A2.5 2.5 0 0 1 3 17.5z" />
+      <path d="M8 11.5h8" />
+      <path d="M8 15h6" />
+    </svg>
+  );
+}
+
+export function Folder({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="M20 20a2 2 0 0 0 2-2V8a2 2 0 0 0-2-2h-7.9a2 2 0 0 1-1.69-.9L9.6 3.9A2 2 0 0 0 7.93 3H4a2 2 0 0 0-2 2v13a2 2 0 0 0 2 2Z" />
+    </svg>
+  );
+}
+
+export function FolderOpen({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} strokeLinecap="round" strokeLinejoin="round" {...props}>
+      <path d="m6 14 1.5-2.9A2 2 0 0 1 9.24 10H20a2 2 0 0 1 1.94 2.5l-1.54 6a2 2 0 0 1-1.95 1.5H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h3.9a2 2 0 0 1 1.69.9l.81 1.2a2 2 0 0 0 1.67.9H18a2 2 0 0 1 2 2v2" />
+    </svg>
+  );
+}
+
+/** Filled closed folder matching Codex's sidebar project folders */
+export function FolderClosedFilled({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 20 20" {...props}>
+      <path d="M6.584 2.835h.695c.143 0 .243 0 .341.006l.254.026a3.01 3.01 0 0 1 1.582.738l.245.24c.112.113.149.15.185.182.267.2.586.32.92.34l.26.003h2.626c.559 0 1.012 0 1.381.026.375.027.713.085 1.032.224l.264.128a3.165 3.165 0 0 1 1.34 1.475l.087.242c.074.246.11.504.13.783.026.366.026.815.026 1.368v4.084c0 .654.001 1.185-.034 1.614-.036.437-.112.827-.297 1.19l-.117.21a3.168 3.168 0 0 1-1.212 1.119l-.138.064c-.324.139-.67.201-1.053.232-.429.035-.96.034-1.614.034H6.585c-.654 0-1.185.001-1.614-.034-.382-.031-.729-.093-1.053-.232l-.138-.064a3.168 3.168 0 0 1-1.212-1.119l-.117-.21c-.185-.363-.261-.753-.297-1.19-.035-.429-.034-.96-.034-1.614V7.3c0-.654-.001-1.185.034-1.614.036-.437.112-.827.297-1.19l.117-.21a3.168 3.168 0 0 1 1.212-1.119l.138-.064c.324-.139.67-.201 1.053-.232.429-.035.96-.034 1.614-.034Zm0 1.33c-.676 0-1.143.001-1.506.031-.266.022-.443.057-.575.104l-.12.052a1.833 1.833 0 0 0-.682.629l-.065.118c-.074.145-.127.341-.156.696-.03.363-.03.83-.03 1.506v5.4c0 .676.001 1.143.03 1.506.029.355.082.551.156.696l.065.118c.164.268.4.486.682.629l.12.052c.132.047.309.082.575.104.363.03.83.031 1.506.031h6.83c.676 0 1.143-.001 1.506-.031.266-.022.443-.057.575-.104l.12-.052c.282-.143.518-.361.682-.629l.065-.118c.074-.145.127-.341.156-.696.03-.363.03-.83.03-1.506V8.615c0-.577-.001-.976-.023-1.287-.021-.305-.06-.476-.114-.603l-.072-.148a1.835 1.835 0 0 0-.754-.753c-.127-.054-.298-.093-.603-.114-.311-.022-.71-.023-1.287-.023h-2.626l-.342-.006a3.004 3.004 0 0 1-1.638-.604l-.197-.162a6.28 6.28 0 0 1-.243-.239l-.186-.182a1.75 1.75 0 0 0-.891-.312l-.142-.015c-.048-.003-.1-.003-.26-.003h-.695Z" />
+    </svg>
+  );
+}
+
+/** Filled folder-plus icon matching Codex's "Add new project" button */
+export function FolderPlusFilled({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 20 20" {...props}>
+      <path d="M15.2041 17.5V15.665H13.3691C13.0019 15.665 12.7041 15.3673 12.7041 15C12.7041 14.6327 13.0019 14.335 13.3691 14.335H15.2041V12.5C15.2041 12.1327 15.5019 11.835 15.8691 11.835C16.2362 11.8352 16.5332 12.1329 16.5332 12.5V14.335H18.3691C18.7362 14.3352 19.0332 14.6329 19.0332 15C19.0332 15.3671 18.7362 15.6648 18.3691 15.665H16.5332V17.5C16.5332 17.8671 16.2362 18.1648 15.8691 18.165C15.5019 18.165 15.2041 17.8673 15.2041 17.5ZM2.12012 12.7002V7.29981C2.12012 6.64581 2.11922 6.1149 2.1543 5.68555C2.19002 5.24867 2.26619 4.85832 2.45117 4.49512L2.56836 4.28516C2.86045 3.80898 3.27979 3.42103 3.78028 3.16602L3.91797 3.10156C4.24192 2.96268 4.5885 2.90039 4.97071 2.86914C5.40006 2.83406 5.93096 2.83496 6.58496 2.83496H7.28028C7.42346 2.83496 7.52305 2.83479 7.6211 2.84082L7.875 2.86719C8.46133 2.95309 9.01189 3.20874 9.45703 3.60547L9.70215 3.84473C9.81425 3.95779 9.85105 3.99455 9.88672 4.02637L9.99805 4.11719C10.2646 4.31741 10.5851 4.43638 10.9199 4.45703L11.1797 4.45996H13.6914C14.2499 4.45996 14.703 4.45958 15.0713 4.48535C15.4458 4.51157 15.7828 4.56683 16.1025 4.70313L16.3662 4.83106C16.9638 5.15706 17.4378 5.67623 17.707 6.30762L17.7939 6.54981C17.868 6.79538 17.904 7.05317 17.9238 7.33203C17.9498 7.69789 17.9502 8.14747 17.9502 8.7002C17.9501 8.87631 17.8803 9.0453 17.7559 9.16992C17.6311 9.29464 17.4615 9.36524 17.2852 9.36524H3.4502V12.7002C3.4502 13.3761 3.45084 13.8434 3.48047 14.2061C3.50947 14.5608 3.56304 14.7568 3.63672 14.9014L3.70215 15.0195C3.86642 15.2873 4.10236 15.505 4.38379 15.6484L4.50391 15.7002C4.63661 15.7476 4.81329 15.783 5.0791 15.8047C5.44174 15.8343 5.90903 15.835 6.58496 15.835H9.40918L9.54395 15.8486C9.84681 15.9108 10.0742 16.1788 10.0742 16.5C10.0742 16.8212 9.84681 17.0892 9.54395 17.1514L9.40918 17.165H6.58496C5.93096 17.165 5.40006 17.1659 4.97071 17.1309C4.5885 17.0996 4.24192 17.0373 3.91797 16.8984L3.78028 16.834C3.27979 16.579 2.86045 16.191 2.56836 15.7148L2.45117 15.5049C2.26619 15.1417 2.19002 14.7513 2.1543 14.3145C2.11922 13.8851 2.12012 13.3542 2.12012 12.7002ZM3.4502 8.03516H16.6172C16.6146 7.79548 16.6098 7.59777 16.5977 7.42676C16.5816 7.20054 16.5552 7.04845 16.5205 6.9336L16.4834 6.8291C16.332 6.47411 16.0655 6.1824 15.7295 5.99903L15.5811 5.92676C15.4545 5.8728 15.2835 5.83385 14.9785 5.8125C14.6674 5.79073 14.2686 5.79004 13.6914 5.79004H11.1797L10.8379 5.78418C10.2426 5.74746 9.67313 5.53663 9.19922 5.18067L9.00196 5.01953C8.92848 4.95403 8.85889 4.88222 8.75781 4.78028L8.57227 4.59863C8.32169 4.37525 8.01175 4.23086 7.68164 4.18262L7.54004 4.16797C7.49225 4.16502 7.43987 4.16504 7.28028 4.16504H6.58496C5.90903 4.16504 5.44174 4.16569 5.0791 4.19531C4.81329 4.21705 4.63661 4.25237 4.50391 4.29981L4.38379 4.35156C4.10236 4.49499 3.86642 4.71271 3.70215 4.98047L3.63672 5.09863C3.56304 5.24324 3.50947 5.43924 3.48047 5.79395C3.45084 6.15659 3.4502 6.62388 3.4502 7.29981V8.03516Z" />
+    </svg>
+  );
+}
+
+/** Filled folder icon matching Codex's sidebar project folders */
+export function FolderFilled({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 20 20" {...props}>
+      <path d="M6.584 2.874a3.01 3.01 0 0 1 1.816.757c.073.064.142.135.243.237.112.113.15.15.187.183.292.26.663.415 1.053.44.049.002.102.002.261.002h2.718c.56 0 1.015 0 1.386.027.377.027.714.086 1.034.226.608.267 1.11.727 1.43 1.31.168.307.256.637.316 1.01.03.181.054.383.077.609h.371a1.915 1.915 0 0 1 1.832 2.475l-1.645 5.367a2.331 2.331 0 0 1-2.229 1.648H4.754c-.61 0-1.15-.23-1.559-.6a3.006 3.006 0 0 1-.847-.933c-.191-.33-.287-.687-.351-1.093-.063-.398-.1-.89-.147-1.499l-.418-5.435c-.052-.683-.096-1.235-.093-1.681.002-.453.05-.858.214-1.237a3.008 3.008 0 0 1 1.365-1.475c.366-.192.766-.27 1.218-.308.444-.036.997-.036 1.682-.036h.427c.144 0 .242 0 .339.006Zm-.66 6.13a.586.586 0 0 0-.559.415l-1.57 5.121a1.002 1.002 0 0 0 .589 1.224c.109.03.244.055.422.071h10.628c.44 0 .828-.288.957-.708l1.645-5.366a.585.585 0 0 0-.56-.756H5.925Zm-.106-4.872c-.706 0-1.198 0-1.579.032-.374.03-.582.087-.734.167a1.744 1.744 0 0 0-.791.855c-.068.157-.11.369-.112.745-.002.382.036.873.09 1.578l.374 4.87 1.028-3.35a1.916 1.916 0 0 1 1.83-1.354h9.909a8.189 8.189 0 0 0-.052-.406c-.05-.304-.107-.476-.178-.606a1.746 1.746 0 0 0-.829-.76c-.135-.059-.312-.1-.618-.123a19.667 19.667 0 0 0-1.294-.023h-2.718c-.143 0-.243 0-.34-.006a3.007 3.007 0 0 1-1.815-.757 6.091 6.091 0 0 1-.243-.237 4.418 4.418 0 0 0-.187-.183 1.745 1.745 0 0 0-1.052-.44c-.05-.002-.103-.002-.262-.002h-.427Z" />
+    </svg>
+  );
+}
+
+/** Filter/funnel icon matching Codex's sidebar filter button */
+export function Filter({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 20 20" {...props}>
+      <path d="M12.5 14.0049C12.8673 14.0049 13.165 14.3027 13.165 14.6699C13.165 15.0372 12.8673 15.335 12.5 15.335H7.5C7.13273 15.335 6.83496 15.0372 6.83496 14.6699C6.83496 14.3027 7.13273 14.0049 7.5 14.0049H12.5Z" />
+      <path d="M15 9.33496C15.3673 9.33496 15.665 9.63273 15.665 10C15.665 10.3673 15.3673 10.665 15 10.665H5C4.63273 10.665 4.33496 10.3673 4.33496 10C4.33496 9.63273 4.63273 9.33496 5 9.33496H15Z" />
+      <path d="M17.5 4.66504C17.8673 4.66504 18.165 4.96281 18.165 5.33008C18.165 5.69735 17.8673 5.99512 17.5 5.99512H2.5C2.13273 5.99512 1.83496 5.69735 1.83496 5.33008C1.83496 4.96281 2.13273 4.66504 2.5 4.66504H17.5Z" />
+    </svg>
+  );
+}
+
+/** Collapse-all icon matching Codex's sidebar collapse button */
+export function CollapseAll({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 20 20" {...props}>
+      <path d="M16.0299 3.0293C16.2896 2.76996 16.7107 2.76988 16.9703 3.0293C17.23 3.28899 17.23 3.711 16.9703 3.9707L13.2731 7.66797H16.9996L17.1344 7.68164C17.4372 7.74375 17.6645 8.01192 17.6647 8.33301C17.6647 8.65421 17.4372 8.92219 17.1344 8.98438L16.9996 8.99805H11.6666C11.2994 8.99801 11.0016 8.70026 11.0016 8.33301V3C11.0016 2.63275 11.2994 2.33499 11.6666 2.33496C12.0339 2.33496 12.3317 2.63273 12.3317 3V6.72754L16.0299 3.0293ZM8.99475 17C8.99475 17.3673 8.69698 17.665 8.32971 17.665C7.96258 17.6649 7.66467 17.3672 7.66467 17V13.2725L3.96741 16.9707C3.70771 17.2304 3.2857 17.2304 3.026 16.9707C2.7663 16.711 2.7663 16.289 3.026 16.0293L6.72424 12.332H2.9967C2.62955 12.332 2.33185 12.0341 2.33167 11.667C2.33167 11.2997 2.62943 11.002 2.9967 11.002H8.32971C8.69698 11.002 8.99475 11.2997 8.99475 11.667V17Z" />
+    </svg>
+  );
+}
+
+/** Expand-all icon — inverse of CollapseAll: arrows point outward from center. */
+export function ExpandAll({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 24 24" {...props}>
+      <path
+        fillRule="evenodd"
+        d="M3 3.75C3 3.336 3.336 3 3.75 3H9a.75.75 0 0 1 0 1.5H5.56l3.97 3.97a.75.75 0 1 1-1.06 1.06L4.5 5.56V9a.75.75 0 0 1-1.5 0V3.75Zm12 0A.75.75 0 0 1 15.75 3H21a.75.75 0 0 1 .75.75V9a.75.75 0 0 1-1.5 0V5.56l-3.97 3.97a.75.75 0 1 1-1.06-1.06l3.97-3.97H15.75A.75.75 0 0 1 15 3.75ZM9.53 15.47a.75.75 0 0 1 0 1.06l-3.97 3.97H9A.75.75 0 0 1 9 22.0H3.75A.75.75 0 0 1 3 21.25V16a.75.75 0 0 1 1.5 0v3.44l3.97-3.97a.75.75 0 0 1 1.06 0Zm4.94 0a.75.75 0 0 1 1.06 0l3.97 3.97V16a.75.75 0 0 1 1.5 0v5.25a.75.75 0 0 1-.75.75h-5.25a.75.75 0 0 1 0-1.5h3.44l-3.97-3.97a.75.75 0 0 1 0-1.06Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function GitCommit({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path d="M13.5013 10.0003C13.5013 8.06653 11.9341 6.49856 10.0003 6.49838C8.06641 6.49838 6.49837 8.06642 6.49837 10.0003C6.49855 11.9341 8.06652 13.5013 10.0003 13.5013C11.934 13.5011 13.5011 11.934 13.5013 10.0003ZM14.8314 10.0003C14.8312 12.6685 12.6685 14.8312 10.0003 14.8314C7.33198 14.8314 5.16847 12.6686 5.16829 10.0003C5.16829 7.33188 7.33187 5.1683 10.0003 5.1683C12.6686 5.16848 14.8314 7.33199 14.8314 10.0003Z" fill="currentColor" />
+      <path d="M5 9.33497C5.36727 9.33497 5.66504 9.63274 5.66504 10C5.66504 10.3673 5.36727 10.665 5 10.665H1.25C0.882731 10.665 0.584961 10.3673 0.584961 10C0.584961 9.63274 0.882731 9.33497 1.25 9.33497H5Z" fill="currentColor" />
+      <path d="M18.75 9.33497C19.1173 9.33497 19.415 9.63274 19.415 10C19.415 10.3673 19.1173 10.665 18.75 10.665H15C14.6327 10.665 14.335 10.3673 14.335 10C14.335 9.63274 14.6327 9.33497 15 9.33497H18.75Z" fill="currentColor" />
+    </svg>
+  );
+}
+
+export function Tree({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 20 20" fill="currentColor" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path
+        d="M15.8 11.535c.367 0 .665.298.665.665v5a.665.665 0 0 1-.665.665h-5a.665.665 0 1 1 0-1.33h3.394l-3.565-3.564a.666.666 0 0 1 .942-.942l3.564 3.565V12.2c0-.367.298-.665.665-.665Zm0-9.4c.367 0 .665.298.665.665v5a.665.665 0 0 1-1.33 0V4.405l-5.128 5.128c-.323.324-.558.565-.842.74a2.668 2.668 0 0 1-.771.319c-.324.078-.662.073-1.12.073H1.93a.665.665 0 1 1 0-1.33h5.345c.52 0 .673-.005.809-.037.136-.033.266-.086.385-.16.12-.072.23-.177.598-.545l5.128-5.128H10.8a.665.665 0 0 1 0-1.33h5Z"
+      />
+    </svg>
+  );
+}
+
+export function GitBranchIcon({ className, ...props }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <circle cx="5.4165" cy="5" r="1.875" stroke="currentColor" strokeWidth="1.33" />
+      <circle cx="5.4165" cy="15" r="1.875" stroke="currentColor" strokeWidth="1.33" />
+      <circle cx="14.5833" cy="5" r="1.875" stroke="currentColor" strokeWidth="1.33" />
+      <path d="M5.4165 6.66664V13.3333" stroke="currentColor" strokeWidth="1.33" strokeLinejoin="round" />
+      <path d="M5.41658 12.5V11.6667C5.41658 10.7462 6.16278 10 7.08325 10H12.9166C13.8371 10 14.5833 9.25381 14.5833 8.33333V7.5" stroke="currentColor" strokeWidth="1.33" strokeLinejoin="round" />
+    </svg>
+  );
+}
